requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-widget-tweaks==1.4.12
listview==0.1.0
Pillow==9.1.1
sqlparse==0.4.2
tzdata==2022.1
